change expert bias to fp32 (merge request !13)
Squash merge branch 'feature/expert_bias_fp32' into 'feature/sgl_mmq'
change expert bias to fp32

+++ python/sglang/srt/models/welmv4.py
@@ -214,7 +214,7 @@ class Qwen2MoeSparseMoeBlock(nn.Module):
     ):
         super().__init__()
         self.tp_size = get_tensor_model_parallel_world_size()
-        self.expert_bias = torch.nn.Parameter(torch.zeros((config.num_experts)))
+        self.expert_bias = torch.nn.Parameter(torch.zeros((config.num_experts), dtype=torch.float32))
         self.layer_id = layer_id
         self.alt_stream = alt_stream
         if self.tp_size > config.num_experts:
